+++ src/plugin/Stdin.c
@@ -125,5 +125,6 @@ Plugin Stdin_Plugin = {
     ARRAY_NULL_INITIALIZER,
     Plugin_get_description,
     Plugin_get_name,
+    NULL,
     Plugin_run,
 };